Restructure for multi-frame messages.
The MATLAB scripts now embed an index in the remaining bits after the
message (previously just 0's).
The Android code now saves message data by in a hashmap of indices to
collections of frame messages, such that for each index we store all
reported messages. When asked to fetch the total message, we iterate
through each index, picking the most popular version of that frame
message and appending the results.